Handle singles + controllers don't need aliases
It is not mandatory anymore to define a section for each service / single you want to use.

If a single do not need any specific section, you can put its complete path in the configuration value.

This commit also introduces the handling od singles and adds support for setup/teardown methods